raspberry-database/requirements.txt
numpy
mysql.connector
schedule